mychatbot-master/requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
pyee==6.0.0
requests==2.22.0
six==1.12.0
slackclient==2.1.0
slackeventsapi==2.1.0
soupsieve==1.9.2
urllib3==1.25.3
Werkzeug==0.15.4
yarl==1.3.0
konlpy==0.5.1